+++ core-backend/sandbox/src/memory.rs
@@ -60,8 +60,8 @@ impl Memory for MemoryWrap {
         (self.0.size() * 65536) as usize
     }
 
-    fn get_wasm_memory_begin_addr(&self) -> usize {
-        unsafe { self.0.get_buff() as usize }
+    fn get_wasm_memory_begin_addr(&self) -> u64 {
+        unsafe { self.0.get_buff() }
     }
 }
